# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WhatsApp OpenWA Rebuild API
  version: 0.1.0
  description: >
    Contract baseline for the active root runtime. This file must be reviewed for
    every backend change and updated in the same work item whenever request,
    response, auth, validation, header, or route behavior changes.
...
servers:
  - url: http://localhost:8192
...
components:
  parameters:
    XForwardedFor:
      name: X-Forwarded-For
      in: header
      required: false
      schema:
        type: string
      description: >
        Optional proxy-forwarded client IP. Admin routes are protected by the
        server-side allowlist and may return `403 Forbidden` when the resolved
        client IP is not allowed.
    CaptureLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CaptureEventType:
      name: eventType
      in: query
      required: true
      schema:
        type: string
  schemas:
    StatusEnvelope:
      type: object
      properties:
        status:
          type: boolean
      required:
        - status
    ErrorEnvelope:
      type: object
      properties:
        status:
          type: boolean
          const: false
        message:
          type: string
      required:
        - status
        - message
    ValidationErrorEnvelope:
      type: object
      properties:
        status:
          type: boolean
          const: false
        errors:
          type: object
          additionalProperties:
            type: string
      required:
        - status
        - errors
...
    CaptureRecord:
      type: object
      properties:
        captureId:
          type: string
        eventType:
          type: string
          nullable: true
        sessionId:
          type: string
          nullable: true
        capturedAt:
          type: string
          format: date-time
      additionalProperties: true
    CaptureListResponse:
      allOf:
        - $ref: '#/components/schemas/StatusEnvelope'
        - type: object
          properties:
            status:
              const: true
            captures:
              type: array
              items:
                $ref: '#/components/schemas/CaptureRecord'
          required:
            - captures
    CaptureSingleResponse:
      allOf:
        - $ref: '#/components/schemas/StatusEnvelope'
        - type: object
          properties:
            status:
              const: true
            capture:
              $ref: '#/components/schemas/CaptureRecord'
          required:
            - capture
    WebhookRegisterRequest:
      type: object
      properties:
        url:
          type: string
          format: uri
        secret:
          type: string
        retryCount:
          type: integer
          minimum: 0
          maximum: 5
        events:
          type: array
          minItems: 1
          items:
            type: string
    WebhookTestRequest:
      type: object
      properties:
        payload:
          type: object
          additionalProperties: true
        headers:
          type: object
          additionalProperties:
            oneOf:
              - type: string
              - type: number
              - type: boolean
      required:
        - payload
    WebhookRegisterResponse:
      allOf:
        - $ref: '#/components/schemas/StatusEnvelope'
        - type: object
          properties:
            status:
              const: true
            created:
              type: boolean
            webhook: {}
          required:
            - created
            - webhook
    WebhookListResponse:
      allOf:
        - $ref: '#/components/schemas/StatusEnvelope'
        - type: object
          properties:
            status:
              const: true
            webhooks: {}
          required:
            - webhooks
    WebhookInboundResponse:
      allOf:
        - $ref: '#/components/schemas/StatusEnvelope'
        - type: object
          properties:
            status:
              const: true
            captured:
              type: object
              properties:
                captureId:
                  type: string
                eventType:
                  type: string
                  nullable: true
                sessionId:
                  type: string
                  nullable: true
                capturedAt:
                  type: string
                  format: date-time
            processed:
              type: object
              properties:
                handled:
                  type: boolean
                commandName:
                  type: string
                  nullable: true
    WebhookTestResponse:
      allOf:
        - $ref: '#/components/schemas/WebhookInboundResponse'
        - type: object
          properties:
            mode:
              type: string
              const: test
  responses:
    Forbidden:
      description: Caller IP is not in the server allowlist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    InternalServerError:
      description: Unexpected server-side failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ValidationFailed:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorEnvelope'
paths:
...
  /channel/webhooks/openwa:
    post:
      tags: [Webhooks]
      summary: Receive inbound OpenWA webhook events
      description: >
        Public inbound endpoint used by OpenWA. This route persists captures,
        attempts event normalization, and routes supported slash commands into
        the root inbound command service.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '202':
          description: Webhook accepted and capture persisted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookInboundResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /channel/webhooks/captures:
    get:
      tags: [Webhooks]
      summary: List recent stored webhook captures
      parameters:
        - $ref: '#/components/parameters/XForwardedFor'
        - $ref: '#/components/parameters/CaptureLimit'
      responses:
        '200':
          description: Capture list returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptureListResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /channel/webhooks/captures/latest:
    get:
      tags: [Webhooks]
      summary: Get the latest capture for a specific event type
      parameters:
        - $ref: '#/components/parameters/XForwardedFor'
        - $ref: '#/components/parameters/CaptureEventType'
      responses:
        '200':
          description: Matching capture returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptureSingleResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: No capture exists for the requested event type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /channel/webhooks:
    get:
      tags: [Webhooks]
      summary: List current session webhook registrations
      parameters:
        - $ref: '#/components/parameters/XForwardedFor'
      responses:
        '200':
          description: Webhook list returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookListResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /channel/webhooks/test:
    post:
      tags: [Webhooks]
      summary: Inject a simulated webhook payload into the root processing pipeline
      description: >
        Admin-only helper endpoint for local verification. The provided payload is
        captured, normalized, and routed through the same processing path used by
        the real OpenWA inbound webhook endpoint.
      parameters:
        - $ref: '#/components/parameters/XForwardedFor'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookTestRequest'
      responses:
        '202':
          description: Test payload accepted and processed through the webhook pipeline
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookTestResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /channel/webhooks/register:
    post:
      tags: [Webhooks]
      summary: Create or ensure the current session webhook registration
      parameters:
        - $ref: '#/components/parameters/XForwardedFor'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookRegisterRequest'
      responses:
        '200':
          description: Existing webhook matched and was reused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookRegisterResponse'
        '201':
          description: New webhook created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookRegisterResponse'
        '400':
          description: No webhook URL available from config or request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalServerError'